+++ Modules/Intent.Modules.Application.MediatR.Behaviours/Templates/EventBusPublishBehaviour/EventBusPublishBehaviourTemplatePartial.cs
@@ -52,7 +52,7 @@ namespace Intent.Modules.Application.MediatR.Behaviours.Templates.EventBusPublis
             ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest.ToRegister($"typeof({ClassName}<,>)")
                 .ForInterface("typeof(IPipelineBehavior<,>)")
                 .WithPriority(4)
-                .ForConcern("Application")
+                .ForConcern("MediatR")
                 .RequiresUsingNamespaces("MediatR")
                 .HasDependency(this));
         }