+++ Modules/Intent.Modules.Application.MediatR.Behaviours/Templates/MongoDbUnitOfWorkBehaviour/MongoDbUnitOfWorkBehaviourTemplatePartial.cs
@@ -73,7 +73,7 @@ namespace Intent.Modules.Application.MediatR.Behaviours.Templates.MongoDbUnitOfW
             ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest.ToRegister($"typeof({ClassName}<,>)")
                 .ForInterface("typeof(IPipelineBehavior<,>)")
                 .WithPriority(4)
-                .ForConcern("Application")
+                .ForConcern("MediatR")
                 .RequiresUsingNamespaces("MediatR")
                 .HasDependency(this));
         }