+++ Modules/Intent.Modules.Application.MediatR.Behaviours/Templates/UnhandledExceptionBehaviour/UnhandledExceptionBehaviourTemplatePartial.cs
@@ -37,7 +37,7 @@ namespace Intent.Modules.Application.MediatR.Behaviours.Templates.UnhandledExcep
             ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest.ToRegister($"typeof({ClassName}<,>)")
                 .ForInterface("typeof(IPipelineBehavior<,>)")
                 .WithPriority(0)
-                .ForConcern("Application")
+                .ForConcern("MediatR")
                 .RequiresUsingNamespaces("MediatR")
                 .HasDependency(this));
         }