+++ Modules/Intent.Modules.Dapr.AspNetCore.StateManagement/Templates/DaprStateStoreUnitOfWorkBehaviour/DaprStateStoreUnitOfWorkBehaviourTemplatePartial.cs
@@ -58,7 +58,7 @@ namespace Intent.Modules.Dapr.AspNetCore.StateManagement.Templates.DaprStateStor
             ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest.ToRegister($"typeof({ClassName}<,>)")
                 .ForInterface("typeof(IPipelineBehavior<,>)")
                 .WithPriority(6)
-                .ForConcern("Application")
+                .ForConcern("MediatR")
                 .RequiresUsingNamespaces("MediatR")
                 .HasDependency(this));
         }