+++ Modules/Intent.Modules.VisualStudio.Projects.Tests/NuGet/SchemeTests/NetFrameworkPackageReferenceSchemeTests.cs
@@ -47,6 +47,7 @@ namespace Intent.Modules.VisualStudio.Projects.Tests.NuGet.SchemeTests
                 project.Content,
                 project.RequestedPackages,
                 project.InstalledPackages,
+                new List<string>(),
                 project.Name,
                 tracing,
                 DependencyVersionOverwriteBehaviorOption.IfNewer);
@@ -84,6 +85,7 @@ namespace Intent.Modules.VisualStudio.Projects.Tests.NuGet.SchemeTests
                 project.Content,
                 project.RequestedPackages,
                 project.InstalledPackages,
+                new List<string>(),
                 project.Name,
                 tracing,
                 DependencyVersionOverwriteBehaviorOption.IfNewer);
@@ -120,6 +122,7 @@ namespace Intent.Modules.VisualStudio.Projects.Tests.NuGet.SchemeTests
                 project.Content,
                 project.RequestedPackages,
                 project.InstalledPackages,
+                new List<string>(),
                 project.Name,
                 tracing,
                 DependencyVersionOverwriteBehaviorOption.IfNewer);