+++ Modules/Intent.Modules.VisualStudio.Projects.Tests/NuGet/SchemeTests/UnsupportedSchemeTests.cs
@@ -40,6 +40,7 @@ namespace Intent.Modules.VisualStudio.Projects.Tests.NuGet.SchemeTests
                 project.Content,
                 project.RequestedPackages,
                 project.InstalledPackages,
+                new List<string>(),
                 project.Name,
                 tracing,
                 DependencyVersionOverwriteBehaviorOption.IfNewer);
@@ -67,9 +68,10 @@ namespace Intent.Modules.VisualStudio.Projects.Tests.NuGet.SchemeTests
                 project.Content,
                 project.RequestedPackages,
                 project.InstalledPackages,
+                new List<string>(),
                 project.Name,
                 tracing,
-                DependencyVersionOverwriteBehaviorOption.IfNewer);
+                DependencyVersionOverwriteBehaviorOption.IfNewer); ;
 
             // Assert
             Assert.Collection(tracing.DebugEntries,