+++ Modules/Intent.Modules.VisualStudio.Projects/FactoryExtensions/NuGet/HelperTypes/INuGetSchemeProcessor.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Xml.Linq;
 using Intent.Engine;
+using Intent.Modules.Common.CSharp.VisualStudio;
 using Intent.Modules.Common.VisualStudio;
 using Intent.Modules.VisualStudio.Projects.Settings;
 
@@ -13,6 +14,7 @@ namespace Intent.Modules.VisualStudio.Projects.NuGet.HelperTypes
         string InstallPackages(string projectContent,
             Dictionary<string, NuGetPackage> requestedPackages,
             Dictionary<string, NuGetPackage> installedPackages,
+            List<string> packagesToRemove,
             string projectName,
             ITracing tracing,
             DependencyVersionOverwriteBehaviorOption dependencyVersionOverwriteBehavior);