+++ Modules/Intent.Modules.VisualStudio.Projects/FactoryExtensions/NuGet/SchemeProcessors/NetFrameworkPackagesDotConfigSchemeProcessor.cs
@@ -62,6 +62,7 @@ namespace Intent.Modules.VisualStudio.Projects.FactoryExtensions.NuGet.SchemePro
         public string InstallPackages(string projectContent,
             Dictionary<string, NuGetPackage> requestedPackages,
             Dictionary<string, NuGetPackage> installedPackages,
+            List<string> toRemovePackages,
             string projectName,
             ITracing tracing,
             DependencyVersionOverwriteBehaviorOption dependencyVersionOverwriteBehavior)